requirements.txt
# MCP SDK
mcp>=1.0.0
mcp-core>=1.0.0
mcp-protocol>=1.0.0

# API and web framework
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
starlette>=0.27.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
requests>=2.31.0
python-multipart>=0.0.6

# Google Cloud dependencies
google-cloud-dataflow>=0.5.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-cloud-logging>=3.6.0

# Apache Beam
apache-beam>=2.49.0
apache-beam[gcp]>=2.49.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.1
httpx>=0.24.1

# Development
black>=23.7.0
isort>=5.12.0
mypy>=1.4.1

# Observability
opentelemetry-api==1.19.0
opentelemetry-sdk==1.19.0
opentelemetry-exporter-otlp==1.19.0
prometheus-client==0.17.1

# Testing
pytest-cov==4.1.0

# Utilities
marshmallow==3.20.1
kubernetes==28.1.0
gitpython==3.1.37 